requirements.txt
### jax stuff
jax
jaxlib

### data
torchtext

### wandb for logging
wandb


### pytorch for compat
pytorch
tvm
